Android: Avoid calling dw_main_sleep() and dw_main_iteration() during expose callbacks.  Expose callbacks are the only ones that run on the UI thread... due to Android implementation restrictions.  Therefore calling any blocking functions such as dw_mutex_lock() will cause a message loop to be run, via dw_main_sleep() or dw_main_iteration(), these use hacky stuctured exception handling methods to return on Android, so to be safe avoid it completely. Call dw_mutex_trylock() instead, and if it can't lock the mutex... call dw_render_redraw() to trigger a new expose event.

+++ cc.h
@@ -39,9 +39,9 @@ typedef struct _instance {
 	void(*Draw)(struct _instance *);
 	void(*Create)(struct _instance *, HWND owner);
 	void(*Update)(struct _instance *, HWND owner);
-	unsigned Flags;
+	unsigned long Flags;
 	int page;
-	int width, height, vsize;
+	unsigned long width, height, vsize;
 	void *custom, *internal;
 	HWND hwnd, other, *hwndDraw;
 	HPIXMAP *pixmap;
